+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 * [BC] Remove Google Bot, Yahoo and BingBot from the crawler's list, because those search engines
 support _escaped_fragment_ and want to ensure people aren't penalized for cloaking
-* The priority the listener is registered has been lowered from 10000 to 1000.
+* Prerender listener now triggers two events: "prerender.pre" and "prerender.post". This allows you to
+cache the response and return it without hitting the Prerender service
+* The priority the listener is registered has been lowered from 10000 to 1000
 
 # 1.1.2